Split up plot functions into separate files
